create Settings component with Calendar and EditProfile. restructured file strucutre. changed auth navbar to include settings.
