+++ client/src/App.js
@@ -5,9 +5,9 @@ import Landing from "./components/landing/Landing";
 import Register from "./components/auth/Register";
 import Login from "./components/auth/Login";
 import Dashboard from "./components/dashboard/Dashboard";
-// import Profile from "./components/dashboard/profile/Profile";
-import CreateProfile from "./components/dashboard/profile/CreateProfile";
-import EditProfile from "./components/dashboard/profile/EditProfile";
+// import Profile from "./components/profile/Profile";
+import CreateProfile from "./components/profile/CreateProfile";
+import Settings from "./components/dashboard/settings/Settings";
 import PrivateRoute from "./components/routing/PrivateRoute";
 // Redux
 import { Provider } from "react-redux";
@@ -52,9 +52,9 @@ const App = () => {
               />
               <PrivateRoute
                 exact
-                path="/edit-profile"
+                path="/settings"
                 access={[REGISTERED, PROFILED, MET]}
-                component={EditProfile}
+                component={Settings}
               />
               <PrivateRoute
                 exact
