+++ client/src/components/dashboard/settings/Calendar.js
@@ -1,0 +1,10 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const Calendar = props => {
+  return <div>Calendar</div>;
+};
+
+Calendar.propTypes = {};
+
+export default Calendar;
