+++ client/src/components/dashboard/settings/Settings.js
@@ -1,0 +1,14 @@
+import React from "react";
+import EditProfile from "../../profile/EditProfile";
+import Calendar from "./Calendar";
+
+const Settings = () => {
+  return (
+    <div>
+      <Calendar />
+      <EditProfile />
+    </div>
+  );
+};
+
+export default Settings;
