+++ client/src/components/profile/CreateProfile.js
@@ -2,7 +2,7 @@ import React, { useState, Fragment } from "react";
 import { withRouter } from "react-router-dom";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { createProfile } from "../../../actions/profile";
+import { createProfile } from "../../actions/profile";
 
 const CreateProfile = ({ createProfile, history }) => {
   const [formData, setFormData] = useState({
